agents/requirements.txt
pyshp
geopandas
shapely
pytest
flake8
pandas
google-adk
folium 
python-multipart
google-cloud-storage
fsspec
gcsfs